Improve sign handling

+++ header/Text/URLString.h
@@ -7,7 +7,7 @@ namespace Text
 	{
 	public:
 		static UTF8Char *GetURLFilePath(UTF8Char *sbuff, const UTF8Char *url);
-		static UTF8Char *GetURLDomain(UTF8Char *sbuff, const UTF8Char *url, Int32 *port);
+		static UTF8Char *GetURLDomain(UTF8Char *sbuff, const UTF8Char *url, UInt16 *port);
 
 		static UTF8Char *GetURIScheme(UTF8Char *sbuff, const UTF8Char *url);
 		static UTF8Char *GetURLHost(UTF8Char *sbuff, const UTF8Char *url);
